+++ pub/jaspr_unrouter/lib/jaspr_unrouter.dart
@@ -5,7 +5,7 @@ export 'src/core/route_data.dart';
 export 'src/core/route_definition.dart';
 export 'src/runtime/unrouter.dart';
 export 'src/runtime/navigation.dart'
-    show UnrouterController, UnrouterBuildContextExtension;
+    show UnrouterBuildContextExtension, UnrouterController, UnrouterScope;
 export 'package:unrouter/unrouter.dart'
     show
         RouteExecutionCancelledException,